requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
djangorestframework-simplejwt>=5.3,<6.0
